[ENHANCEMENT] Fix delegated auth error route & move auth check modifying the router
