+++ ui/app/src/Router.tsx
@@ -15,26 +15,31 @@ import { SnackbarProvider } from '@perses-dev/components';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { lazy, ReactElement, Suspense } from 'react';
 import { CookiesProvider } from 'react-cookie';
-import { createBrowserRouter, Navigate, Outlet, RouterProvider } from 'react-router-dom';
+import { createBrowserRouter, Navigate, Outlet, RouterProvider, useLocation } from 'react-router-dom';
 import { QueryParamProvider } from 'use-query-params';
 import { ReactRouter6Adapter } from 'use-query-params/adapters/react-router-6';
 import { PersesLoader } from './components/PersesLoader';
 import { AuthorizationProvider } from './context/Authorization';
-import { ConfigContextProvider, useIsEphemeralDashboardEnabled, useIsExplorerEnabled } from './context/Config';
+import {
+  ConfigContextProvider,
+  useIsAuthEnabled,
+  useIsDelegatedAuthnProviderEnabled,
+  useIsEphemeralDashboardEnabled,
+  useIsExplorerEnabled,
+} from './context/Config';
 import { DarkModeContextProvider } from './context/DarkMode';
 import { NavHistoryProvider } from './context/DashboardNavHistory';
 import {
   AdminRoute,
   ConfigRoute,
-  ExploreRoute,
   DelegatedAuthnErrorRoute,
+  ExploreRoute,
   ImportRoute,
   ProfileRoute,
   ProjectRoute,
   SignInRoute,
   SignUpRoute,
 } from './model/route';
-import { RequireAuth, RequireAuthEnabled } from './views/auth/RequireAuth';
 import SignInView from './views/auth/SignInView';
 import SignUpView from './views/auth/SignUpView';
 import DelegatedAuthnErrorView from './views/auth/DelegatedAuthnErrorView';
@@ -42,6 +47,7 @@ import HomeView from './views/home/HomeView';
 // Default route is eagerly loaded
 import App from './App';
 import { PERSES_APP_CONFIG } from './config';
+import { buildRedirectQueryString, useIsLoggedIn } from './model/auth/auth-client';
 
 // Other routes are lazy-loaded for code-splitting
 const ImportView = lazy(() => import('./views/import/ImportView'));
@@ -187,6 +193,38 @@ function RequireExplorerEnabled(): ReactElement {
 function RequireEphemeralDashboardEnabled(): ReactElement {
   const isEphemeralDashboardEnabled = useIsEphemeralDashboardEnabled();
   if (!isEphemeralDashboardEnabled) {
+    return <Navigate to="/" replace />;
+  }
+  return <Outlet />;
+}
+
+/**
+ * This component aims to redirect the user to the SignIn page if not logged in
+ * or to the delegated authn error page if there is an error with the delegated authn.
+ * Otherwise, it just loads the underlying component(s) defined as children.
+ * This is leveraging the following mechanism:
+ * https://reactrouter.com/en/main/upgrading/v5#refactor-custom-routes
+ * https://gist.github.com/mjackson/d54b40a094277b7afdd6b81f51a0393f
+ * @constructor
+ */
+function RequireAuth(): ReactElement | null {
+  const isDelegatedAuthnProviderEnabled = useIsDelegatedAuthnProviderEnabled();
+  const isAuthEnabled = useIsAuthEnabled();
+  const isLoggedIn = useIsLoggedIn();
+  const location = useLocation();
+  if (!isAuthEnabled || isLoggedIn) {
+    return <Outlet />;
+  }
+  let to = isDelegatedAuthnProviderEnabled ? DelegatedAuthnErrorRoute : SignInRoute;
+  if (location.pathname !== '' && location.pathname !== '/') {
+    to += `?${buildRedirectQueryString(location.pathname + location.search)}`;
+  }
+  return <Navigate to={to} />;
+}
+
+function RequireAuthEnabled(): ReactElement {
+  const isAuthEnabled = useIsAuthEnabled();
+  if (!isAuthEnabled) {
     return <Navigate to="/" replace />;
   }
   return <Outlet />;
